+++ examples/light-switch-app/silabs/BUILD.gn
@@ -107,9 +107,6 @@ silabs_executable("light_switch_app") {
   sources = [
     "${examples_common_plat_dir}/main.cpp",
     "src/AppTask.cpp",
-    "src/BindingHandler.cpp",
-    "src/DataModelCallbacks.cpp",
-    "src/LightSwitchMgr.cpp",
   ]
 
   deps = [
