+++ examples/light-switch-app/silabs/src/AppTask.cpp
@@ -24,19 +24,33 @@
 #include "AppTask.h"
 #include "AppConfig.h"
 #include "AppEvent.h"
-#include "BindingHandler.h"
 #include "LEDWidget.h"
-#include "LightSwitchMgr.h"
+
+#if defined(ENABLE_CHIP_SHELL)
+#include "ShellCommands.h"
+#endif // defined(ENABLE_CHIP_SHELL)
+
 #ifdef DISPLAY_ENABLED
 #include "lcd.h"
 #ifdef QR_CODE_ENABLED
 #include "qrcodegen.h"
 #endif // QR_CODE_ENABLED
 #endif // DISPLAY_ENABLED
+
+#include <app-common/zap-generated/attributes/Accessors.h>
+#include <app-common/zap-generated/cluster-objects.h>
+#include <app-common/zap-generated/ids/Attributes.h>
+#include <app-common/zap-generated/ids/Clusters.h>
+#include <app/CommandSender.h>
+#include <app/ConcreteAttributePath.h>
+#include <app/clusters/bindings/BindingManager.h>
+#include <app/clusters/switch-server/switch-server.h>
 #include <app/server/Server.h>
 #include <app/util/attribute-storage.h>
 #include <assert.h>
+#include <controller/InvokeInteraction.h>
 #include <lib/support/CodeUtils.h>
+#include <lib/support/logging/CHIPLogging.h>
 #include <platform/CHIPDeviceLayer.h>
 #include <platform/silabs/platformAbstraction/SilabsPlatform.h>
 #include <setup_payload/OnboardingCodesUtil.h>
@@ -56,11 +70,12 @@ constexpr chip::EndpointId kGenericSwitchEndpoint = 2;
 
 using namespace chip;
 using namespace chip::app;
+using namespace chip::app::Clusters;
+using namespace chip::app::Clusters::LevelControl;
 using namespace ::chip::DeviceLayer;
 using namespace ::chip::DeviceLayer::Silabs;
 
 using namespace chip::TLV;
-using namespace ::chip::DeviceLayer;
 
 /**********************************************************
  * AppTask Definitions
@@ -78,10 +93,10 @@ CHIP_ERROR AppTask::AppInit()
     CHIP_ERROR err = CHIP_NO_ERROR;
     chip::DeviceLayer::Silabs::GetPlatform().SetButtonsCb(AppTask::ButtonEventHandler);
 
-    err = LightSwitchMgr::GetInstance().Init(kLightSwitchEndpoint, kGenericSwitchEndpoint);
+    err = InitLightSwitch(kLightSwitchEndpoint, kGenericSwitchEndpoint);
     if (err != CHIP_NO_ERROR)
     {
-        SILABS_LOG("LightSwitchMgr Init failed!");
+        SILABS_LOG("InitLightSwitch failed!");
         appError(err);
     }
 
@@ -92,7 +107,6 @@ CHIP_ERROR AppTask::AppInit()
 
 void AppTask::Timer::Start()
 {
-    // Starts or restarts the function timer
     osStatus_t status = osTimerStart(mHandler, pdMS_TO_TICKS(LONG_PRESS_TIMEOUT_MS));
     if (status != osOK)
     {
@@ -236,7 +250,7 @@ void AppTask::AppEventHandler(AppEvent * aEvent)
         if (actionButtonPressed)
         {
             actionButtonSuppressed = true;
-            LightSwitchMgr::GetInstance().changeStepMode();
+            sAppTask.changeStepMode();
         }
         else
         {
@@ -265,11 +279,11 @@ void AppTask::AppEventHandler(AppEvent * aEvent)
     }
     case AppEvent::kEventType_ActionButtonPressed:
         actionButtonPressed = true;
-        LightSwitchMgr::GetInstance().SwitchActionEventHandler(aEvent->Type);
+        sAppTask.SwitchActionEventHandler(aEvent->Type);
         if (functionButtonPressed)
         {
             actionButtonSuppressed = true;
-            LightSwitchMgr::GetInstance().changeStepMode();
+            sAppTask.changeStepMode();
         }
         else if (sAppTask.longPressTimer)
         {
@@ -289,14 +303,590 @@ void AppTask::AppEventHandler(AppEvent * aEvent)
         else
         {
             aEvent->Type = AppEvent::kEventType_TriggerToggle;
-            LightSwitchMgr::GetInstance().SwitchActionEventHandler(aEvent->Type);
+            sAppTask.SwitchActionEventHandler(aEvent->Type);
         }
         aEvent->Type = AppEvent::kEventType_ActionButtonReleased;
-        LightSwitchMgr::GetInstance().SwitchActionEventHandler(aEvent->Type);
+        sAppTask.SwitchActionEventHandler(aEvent->Type);
         break;
     case AppEvent::kEventType_TriggerLevelControlAction:
-        LightSwitchMgr::GetInstance().SwitchActionEventHandler(aEvent->Type);
+        sAppTask.SwitchActionEventHandler(aEvent->Type);
     default:
         break;
     }
+}
+
+CHIP_ERROR AppTask::InitLightSwitch(EndpointId lightSwitchEndpoint, EndpointId genericSwitchEndpoint)
+{
+    VerifyOrReturnError(lightSwitchEndpoint != kInvalidEndpointId, CHIP_ERROR_INVALID_ARGUMENT);
+    VerifyOrReturnError(genericSwitchEndpoint != kInvalidEndpointId, CHIP_ERROR_INVALID_ARGUMENT);
+
+    mLightSwitchEndpoint   = lightSwitchEndpoint;
+    mGenericSwitchEndpoint = genericSwitchEndpoint;
+
+    CHIP_ERROR error = InitBindingHandler();
+    if (error != CHIP_NO_ERROR)
+    {
+        SILABS_LOG("InitBindingHandler() failed!");
+        appError(error);
+    }
+
+#if defined(ENABLE_CHIP_SHELL)
+    LightSwitchCommands::RegisterSwitchCommands();
+#endif // defined(ENABLE_CHIP_SHELL)
+
+    return error;
+}
+
+void AppTask::GenericSwitchOnInitialPress()
+{
+    GenericSwitchEventData * data = Platform::New<GenericSwitchEventData>();
+
+    data->endpoint = mGenericSwitchEndpoint;
+    data->event    = Switch::Events::InitialPress::Id;
+
+    TEMPORARY_RETURN_IGNORED DeviceLayer::PlatformMgr().ScheduleWork(GenericSwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
+}
+
+void AppTask::GenericSwitchOnShortRelease()
+{
+    GenericSwitchEventData * data = Platform::New<GenericSwitchEventData>();
+
+    data->endpoint = mGenericSwitchEndpoint;
+    data->event    = Switch::Events::ShortRelease::Id;
+
+    TEMPORARY_RETURN_IGNORED DeviceLayer::PlatformMgr().ScheduleWork(GenericSwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
+}
+
+StepModeEnum AppTask::getStepMode()
+{
+    return mStepDirection;
+}
+
+void AppTask::changeStepMode()
+{
+    mStepDirection = (mStepDirection == StepModeEnum::kUp) ? StepModeEnum::kDown : StepModeEnum::kUp;
+    ChipLogProgress(AppServer, "Step direction changed. Current Step Direction : %s",
+                    ((getStepMode() == StepModeEnum::kUp) ? "kUp" : "kDown"));
+}
+
+void AppTask::TriggerLightSwitchAction(AppTask::LightSwitchAction action, bool isGroupCommand)
+{
+    BindingCommandData * data = Platform::New<BindingCommandData>();
+
+    data->clusterId = Clusters::OnOff::Id;
+    data->isGroup   = isGroupCommand;
+
+    switch (action)
+    {
+    case AppTask::LightSwitchAction::Toggle:
+        data->commandId = OnOff::Commands::Toggle::Id;
+        break;
+
+    case AppTask::LightSwitchAction::On:
+        data->commandId = OnOff::Commands::On::Id;
+        break;
+
+    case AppTask::LightSwitchAction::Off:
+        data->commandId = OnOff::Commands::Off::Id;
+        break;
+
+    default:
+        Platform::Delete(data);
+        return;
+    }
+
+    TEMPORARY_RETURN_IGNORED DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
+}
+
+void AppTask::TriggerLevelControlAction(StepModeEnum stepMode, bool isGroupCommand)
+{
+    BindingCommandData * data = Platform::New<BindingCommandData>();
+
+    data->clusterId = Clusters::LevelControl::Id;
+    data->isGroup   = isGroupCommand;
+    data->commandId = LevelControl::Commands::StepWithOnOff::Id;
+
+    BindingCommandData::Step stepData{
+        .stepMode       = stepMode,
+        .stepSize       = kStepCommand.stepSize,
+        .transitionTime = kStepCommand.transitionTime
+    };
+    stepData.optionsMask.Set(kStepCommand.optionsMask);
+    stepData.optionsOverride.Set(kStepCommand.optionsOverride);
+    data->commandData = stepData;
+
+    TEMPORARY_RETURN_IGNORED DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
+}
+
+void AppTask::GenericSwitchWorkerFunction(intptr_t context)
+{
+    GenericSwitchEventData * data = reinterpret_cast<GenericSwitchEventData *>(context);
+
+    switch (data->event)
+    {
+    case Switch::Events::InitialPress::Id: {
+        uint8_t currentPosition = 1;
+
+        auto switchCluster = Clusters::Switch::FindClusterOnEndpoint(data->endpoint);
+        if (switchCluster != nullptr)
+        {
+            if (switchCluster->SetCurrentPosition(currentPosition) == CHIP_NO_ERROR)
+            {
+                RETURN_SAFELY_IGNORED switchCluster->OnInitialPress(currentPosition);
+            }
+        }
+        break;
+    }
+    case Switch::Events::ShortRelease::Id: {
+        uint8_t previousPosition = 1;
+        uint8_t currentPosition  = 0;
+
+        auto switchCluster = Clusters::Switch::FindClusterOnEndpoint(data->endpoint);
+        if (switchCluster != nullptr)
+        {
+            if (switchCluster->SetCurrentPosition(currentPosition) == CHIP_NO_ERROR)
+            {
+                RETURN_SAFELY_IGNORED switchCluster->OnShortRelease(previousPosition);
+            }
+        }
+        break;
+    }
+    default:
+        break;
+    }
+
+    Platform::Delete(data);
+}
+
+void AppTask::SwitchActionEventHandler(uint16_t eventType)
+{
+    switch (eventType)
+    {
+    case AppEvent::kEventType_ActionButtonPressed:
+        GenericSwitchOnInitialPress();
+        break;
+    case AppEvent::kEventType_ActionButtonReleased:
+        GenericSwitchOnShortRelease();
+        break;
+    case AppEvent::kEventType_TriggerLevelControlAction:
+        TriggerLevelControlAction(getStepMode());
+        break;
+    case AppEvent::kEventType_TriggerToggle:
+        TriggerLightSwitchAction(AppTask::LightSwitchAction::Toggle);
+        break;
+    default:
+        break;
+    }
+}
+
+void AppTask::ProcessOnOffUnicastBindingCommand(CommandId commandId, const Binding::TableEntry & binding,
+                                                Messaging::ExchangeManager * exchangeMgr, const SessionHandle & sessionHandle)
+{
+    auto onSuccess = [](const ConcreteCommandPath & commandPath, const StatusIB & status, const auto & dataResponse) {
+        ChipLogProgress(NotSpecified, "OnOff command succeeds");
+    };
+
+    auto onFailure = [](CHIP_ERROR error) {
+        ChipLogError(NotSpecified, "OnOff command failed: %" CHIP_ERROR_FORMAT, error.Format());
+    };
+
+    switch (commandId)
+    {
+    case Clusters::OnOff::Commands::Toggle::Id: {
+        Clusters::OnOff::Commands::Toggle::Type toggleCommand;
+        TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(exchangeMgr, sessionHandle, binding.remote, toggleCommand,
+                                                                  onSuccess, onFailure);
+        break;
+    }
+    case Clusters::OnOff::Commands::On::Id: {
+        Clusters::OnOff::Commands::On::Type onCommand;
+        TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(exchangeMgr, sessionHandle, binding.remote, onCommand, onSuccess,
+                                                                  onFailure);
+        break;
+    }
+    case Clusters::OnOff::Commands::Off::Id: {
+        Clusters::OnOff::Commands::Off::Type offCommand;
+        TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(exchangeMgr, sessionHandle, binding.remote, offCommand, onSuccess,
+                                                                  onFailure);
+        break;
+    }
+    }
+}
+
+void AppTask::ProcessOnOffGroupBindingCommand(CommandId commandId, const Binding::TableEntry & binding)
+{
+    Messaging::ExchangeManager & exchangeMgr = Server::GetInstance().GetExchangeManager();
+
+    switch (commandId)
+    {
+    case Clusters::OnOff::Commands::Toggle::Id: {
+        Clusters::OnOff::Commands::Toggle::Type toggleCommand;
+        TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                       toggleCommand);
+        break;
+    }
+    case Clusters::OnOff::Commands::On::Id: {
+        Clusters::OnOff::Commands::On::Type onCommand;
+        TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                       onCommand);
+        break;
+    }
+    case Clusters::OnOff::Commands::Off::Id: {
+        Clusters::OnOff::Commands::Off::Type offCommand;
+        TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                       offCommand);
+        break;
+    }
+    }
+}
+
+void AppTask::ProcessLevelControlUnicastBindingCommand(BindingCommandData * data, const Binding::TableEntry & binding,
+                                                       OperationalDeviceProxy * peer_device)
+{
+    auto onSuccess = [](const ConcreteCommandPath & commandPath, const StatusIB & status, const auto & dataResponse) {
+        ChipLogProgress(NotSpecified, "LevelControl command succeeds");
+    };
+
+    auto onFailure = [](CHIP_ERROR error) {
+        ChipLogError(NotSpecified, "LevelControl command failed: %" CHIP_ERROR_FORMAT, error.Format());
+    };
+
+    VerifyOrDie(peer_device != nullptr && peer_device->ConnectionReady());
+
+    switch (data->commandId)
+    {
+    case Clusters::LevelControl::Commands::MoveToLevel::Id: {
+        Clusters::LevelControl::Commands::MoveToLevel::Type moveToLevelCommand;
+        if (auto moveToLevel = std::get_if<BindingCommandData::MoveToLevel>(&data->commandData))
+        {
+            moveToLevelCommand.level           = moveToLevel->level;
+            moveToLevelCommand.transitionTime  = moveToLevel->transitionTime;
+            moveToLevelCommand.optionsMask     = moveToLevel->optionsMask;
+            moveToLevelCommand.optionsOverride = moveToLevel->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(peer_device->GetExchangeManager(),
+                                                                      peer_device->GetSecureSession().Value(), binding.remote,
+                                                                      moveToLevelCommand, onSuccess, onFailure);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::Move::Id: {
+        Clusters::LevelControl::Commands::Move::Type moveCommand;
+        if (auto move = std::get_if<BindingCommandData::Move>(&data->commandData))
+        {
+            moveCommand.moveMode        = move->moveMode;
+            moveCommand.rate            = move->rate;
+            moveCommand.optionsMask     = move->optionsMask;
+            moveCommand.optionsOverride = move->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(peer_device->GetExchangeManager(),
+                                                                      peer_device->GetSecureSession().Value(), binding.remote,
+                                                                      moveCommand, onSuccess, onFailure);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::Step::Id: {
+        Clusters::LevelControl::Commands::Step::Type stepCommand;
+        if (auto step = std::get_if<BindingCommandData::Step>(&data->commandData))
+        {
+            stepCommand.stepMode        = step->stepMode;
+            stepCommand.stepSize        = step->stepSize;
+            stepCommand.transitionTime  = step->transitionTime;
+            stepCommand.optionsMask     = step->optionsMask;
+            stepCommand.optionsOverride = step->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(peer_device->GetExchangeManager(),
+                                                                      peer_device->GetSecureSession().Value(), binding.remote,
+                                                                      stepCommand, onSuccess, onFailure);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::Stop::Id: {
+        Clusters::LevelControl::Commands::Stop::Type stopCommand;
+        if (auto stop = std::get_if<BindingCommandData::Stop>(&data->commandData))
+        {
+            stopCommand.optionsMask     = stop->optionsMask;
+            stopCommand.optionsOverride = stop->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(peer_device->GetExchangeManager(),
+                                                                      peer_device->GetSecureSession().Value(), binding.remote,
+                                                                      stopCommand, onSuccess, onFailure);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::MoveToLevelWithOnOff::Id: {
+        Clusters::LevelControl::Commands::MoveToLevelWithOnOff::Type moveToLevelWithOnOffCommand;
+        if (auto moveToLevel = std::get_if<BindingCommandData::MoveToLevel>(&data->commandData))
+        {
+            moveToLevelWithOnOffCommand.level           = moveToLevel->level;
+            moveToLevelWithOnOffCommand.transitionTime  = moveToLevel->transitionTime;
+            moveToLevelWithOnOffCommand.optionsMask     = moveToLevel->optionsMask;
+            moveToLevelWithOnOffCommand.optionsOverride = moveToLevel->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(peer_device->GetExchangeManager(),
+                                                                      peer_device->GetSecureSession().Value(), binding.remote,
+                                                                      moveToLevelWithOnOffCommand, onSuccess, onFailure);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::MoveWithOnOff::Id: {
+        Clusters::LevelControl::Commands::MoveWithOnOff::Type moveWithOnOffCommand;
+        if (auto move = std::get_if<BindingCommandData::Move>(&data->commandData))
+        {
+            moveWithOnOffCommand.moveMode        = move->moveMode;
+            moveWithOnOffCommand.rate            = move->rate;
+            moveWithOnOffCommand.optionsMask     = move->optionsMask;
+            moveWithOnOffCommand.optionsOverride = move->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(peer_device->GetExchangeManager(),
+                                                                      peer_device->GetSecureSession().Value(), binding.remote,
+                                                                      moveWithOnOffCommand, onSuccess, onFailure);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::StepWithOnOff::Id: {
+        Clusters::LevelControl::Commands::StepWithOnOff::Type stepWithOnOffCommand;
+        if (auto step = std::get_if<BindingCommandData::Step>(&data->commandData))
+        {
+            stepWithOnOffCommand.stepMode        = step->stepMode;
+            stepWithOnOffCommand.stepSize        = step->stepSize;
+            stepWithOnOffCommand.transitionTime  = step->transitionTime;
+            stepWithOnOffCommand.optionsMask     = step->optionsMask;
+            stepWithOnOffCommand.optionsOverride = step->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(peer_device->GetExchangeManager(),
+                                                                      peer_device->GetSecureSession().Value(), binding.remote,
+                                                                      stepWithOnOffCommand, onSuccess, onFailure);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::StopWithOnOff::Id: {
+        Clusters::LevelControl::Commands::StopWithOnOff::Type stopWithOnOffCommand;
+        if (auto stop = std::get_if<BindingCommandData::Stop>(&data->commandData))
+        {
+            stopWithOnOffCommand.optionsMask     = stop->optionsMask;
+            stopWithOnOffCommand.optionsOverride = stop->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeCommandRequest(peer_device->GetExchangeManager(),
+                                                                      peer_device->GetSecureSession().Value(), binding.remote,
+                                                                      stopWithOnOffCommand, onSuccess, onFailure);
+        }
+        break;
+    }
+    default:
+        break;
+    }
+}
+
+void AppTask::ProcessLevelControlGroupBindingCommand(BindingCommandData * data, const Binding::TableEntry & binding)
+{
+    Messaging::ExchangeManager & exchangeMgr = Server::GetInstance().GetExchangeManager();
+
+    switch (data->commandId)
+    {
+    case Clusters::LevelControl::Commands::MoveToLevel::Id: {
+        Clusters::LevelControl::Commands::MoveToLevel::Type moveToLevelCommand;
+        if (auto moveToLevel = std::get_if<BindingCommandData::MoveToLevel>(&data->commandData))
+        {
+            moveToLevelCommand.level           = moveToLevel->level;
+            moveToLevelCommand.transitionTime  = moveToLevel->transitionTime;
+            moveToLevelCommand.optionsMask     = moveToLevel->optionsMask;
+            moveToLevelCommand.optionsOverride = moveToLevel->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                           moveToLevelCommand);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::Move::Id: {
+        Clusters::LevelControl::Commands::Move::Type moveCommand;
+        if (auto move = std::get_if<BindingCommandData::Move>(&data->commandData))
+        {
+            moveCommand.moveMode        = move->moveMode;
+            moveCommand.rate            = move->rate;
+            moveCommand.optionsMask     = move->optionsMask;
+            moveCommand.optionsOverride = move->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                           moveCommand);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::Step::Id: {
+        Clusters::LevelControl::Commands::Step::Type stepCommand;
+        if (auto step = std::get_if<BindingCommandData::Step>(&data->commandData))
+        {
+            stepCommand.stepMode        = step->stepMode;
+            stepCommand.stepSize        = step->stepSize;
+            stepCommand.transitionTime  = step->transitionTime;
+            stepCommand.optionsMask     = step->optionsMask;
+            stepCommand.optionsOverride = step->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                           stepCommand);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::Stop::Id: {
+        Clusters::LevelControl::Commands::Stop::Type stopCommand;
+        if (auto stop = std::get_if<BindingCommandData::Stop>(&data->commandData))
+        {
+            stopCommand.optionsMask     = stop->optionsMask;
+            stopCommand.optionsOverride = stop->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                           stopCommand);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::MoveToLevelWithOnOff::Id: {
+        Clusters::LevelControl::Commands::MoveToLevelWithOnOff::Type moveToLevelWithOnOffCommand;
+        if (auto moveToLevel = std::get_if<BindingCommandData::MoveToLevel>(&data->commandData))
+        {
+            moveToLevelWithOnOffCommand.level           = moveToLevel->level;
+            moveToLevelWithOnOffCommand.transitionTime  = moveToLevel->transitionTime;
+            moveToLevelWithOnOffCommand.optionsMask     = moveToLevel->optionsMask;
+            moveToLevelWithOnOffCommand.optionsOverride = moveToLevel->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                           moveToLevelWithOnOffCommand);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::MoveWithOnOff::Id: {
+        Clusters::LevelControl::Commands::MoveWithOnOff::Type moveWithOnOffCommand;
+        if (auto move = std::get_if<BindingCommandData::Move>(&data->commandData))
+        {
+            moveWithOnOffCommand.moveMode        = move->moveMode;
+            moveWithOnOffCommand.rate            = move->rate;
+            moveWithOnOffCommand.optionsMask     = move->optionsMask;
+            moveWithOnOffCommand.optionsOverride = move->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                           moveWithOnOffCommand);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::StepWithOnOff::Id: {
+        Clusters::LevelControl::Commands::StepWithOnOff::Type stepWithOnOffCommand;
+        if (auto step = std::get_if<BindingCommandData::Step>(&data->commandData))
+        {
+            stepWithOnOffCommand.stepMode        = step->stepMode;
+            stepWithOnOffCommand.stepSize        = step->stepSize;
+            stepWithOnOffCommand.transitionTime  = step->transitionTime;
+            stepWithOnOffCommand.optionsMask     = step->optionsMask;
+            stepWithOnOffCommand.optionsOverride = step->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                           stepWithOnOffCommand);
+        }
+        break;
+    }
+    case Clusters::LevelControl::Commands::StopWithOnOff::Id: {
+        Clusters::LevelControl::Commands::StopWithOnOff::Type stopWithOnOffCommand;
+        if (auto stop = std::get_if<BindingCommandData::Stop>(&data->commandData))
+        {
+            stopWithOnOffCommand.optionsMask     = stop->optionsMask;
+            stopWithOnOffCommand.optionsOverride = stop->optionsOverride;
+            TEMPORARY_RETURN_IGNORED Controller::InvokeGroupCommandRequest(&exchangeMgr, binding.fabricIndex, binding.groupId,
+                                                                           stopWithOnOffCommand);
+        }
+        break;
+    }
+    default:
+        break;
+    }
+}
+
+void AppTask::LightSwitchChangedHandler(const Binding::TableEntry & binding, OperationalDeviceProxy * peer_device, void * context)
+{
+    VerifyOrReturn(context != nullptr, ChipLogError(NotSpecified, "OnDeviceConnectedFn: context is null"));
+    BindingCommandData * data = static_cast<BindingCommandData *>(context);
+
+    if (binding.type == Binding::MATTER_MULTICAST_BINDING && data->isGroup)
+    {
+        switch (data->clusterId)
+        {
+        case Clusters::OnOff::Id:
+            ProcessOnOffGroupBindingCommand(data->commandId, binding);
+            break;
+        case Clusters::LevelControl::Id:
+            ProcessLevelControlGroupBindingCommand(data, binding);
+            break;
+        }
+    }
+    else if (binding.type == Binding::MATTER_UNICAST_BINDING && !data->isGroup)
+    {
+        switch (data->clusterId)
+        {
+        case Clusters::OnOff::Id:
+            VerifyOrDie(peer_device != nullptr && peer_device->ConnectionReady());
+            ProcessOnOffUnicastBindingCommand(data->commandId, binding, peer_device->GetExchangeManager(),
+                                              peer_device->GetSecureSession().Value());
+            break;
+        case Clusters::LevelControl::Id:
+            ProcessLevelControlUnicastBindingCommand(data, binding, peer_device);
+            break;
+        }
+    }
+}
+
+void AppTask::LightSwitchContextReleaseHandler(void * context)
+{
+    VerifyOrReturn(context != nullptr, ChipLogError(NotSpecified, "LightSwitchContextReleaseHandler: context is null"));
+    Platform::Delete(static_cast<BindingCommandData *>(context));
+}
+
+void AppTask::InitBindingHandlerInternal(intptr_t arg)
+{
+    auto & server = chip::Server::GetInstance();
+    LogErrorOnFailure(Binding::Manager::GetInstance().Init(
+        { &server.GetFabricTable(), server.GetCASESessionManager(), &server.GetPersistentStorage() }));
+    Binding::Manager::GetInstance().RegisterBoundDeviceChangedHandler(LightSwitchChangedHandler);
+    Binding::Manager::GetInstance().RegisterBoundDeviceContextReleaseHandler(LightSwitchContextReleaseHandler);
+}
+
+void AppTask::SwitchWorkerFunction(intptr_t context)
+{
+    VerifyOrReturn(context != 0, ChipLogError(NotSpecified, "SwitchWorkerFunction - Invalid work data"));
+
+    BindingCommandData * data = reinterpret_cast<BindingCommandData *>(context);
+    CHIP_ERROR err            = Binding::Manager::GetInstance().NotifyBoundClusterChanged(data->localEndpointId, data->clusterId,
+                                                                                          static_cast<void *>(data));
+
+    if (err == CHIP_ERROR_INCORRECT_STATE || err == CHIP_ERROR_HANDLER_NOT_SET || err == CHIP_ERROR_NO_MEMORY)
+    {
+        Platform::Delete(data);
+    }
+}
+
+void AppTask::BindingWorkerFunction(intptr_t context)
+{
+    VerifyOrReturn(context != 0, ChipLogError(NotSpecified, "BindingWorkerFunction - Invalid work data"));
+
+    Binding::TableEntry * entry = reinterpret_cast<Binding::TableEntry *>(context);
+    TEMPORARY_RETURN_IGNORED AddBindingEntry(*entry);
+
+    Platform::Delete(entry);
+}
+
+CHIP_ERROR AppTask::InitBindingHandler()
+{
+    TEMPORARY_RETURN_IGNORED chip::DeviceLayer::PlatformMgr().ScheduleWork(InitBindingHandlerInternal);
+    return CHIP_NO_ERROR;
+}
+
+void AppTask::PostAttributeChangeCallback(const chip::app::ConcreteAttributePath & attributePath, uint8_t type, uint16_t size,
+                                          uint8_t * value)
+{
+    ClusterId clusterId                      = attributePath.mClusterId;
+    [[maybe_unused]] AttributeId attributeId = attributePath.mAttributeId;
+    ChipLogDetail(Zcl, "Cluster callback: " ChipLogFormatMEI, ChipLogValueMEI(clusterId));
+
+    if (clusterId == Identify::Id)
+    {
+        ChipLogProgress(Zcl, "Identify attribute ID: " ChipLogFormatMEI " Type: %u Value: %u, length %u",
+                        ChipLogValueMEI(attributeId), type, *value, size);
+    }
+}
+
+void AppTask::OnOffClusterInitCallback(chip::EndpointId endpoint)
+{
+    // TODO: implement any additional Cluster Server init actions
+}
+
+void MatterPostAttributeChangeCallback(const chip::app::ConcreteAttributePath & attributePath, uint8_t type, uint16_t size,
+                                       uint8_t * value)
+{
+    AppTask::PostAttributeChangeCallback(attributePath, type, size, value);
+}
+
+void emberAfOnOffClusterInitCallback(EndpointId endpoint)
+{
+    AppTask::OnOffClusterInitCallback(endpoint);
 }
